[la.container] let dim return unsigned int

+++ dune/pymor/la/container/dunedynamic.hh
@@ -74,7 +74,7 @@ public:
   }
 
   //! \copydoc Dune::Pymor::LA::VectorInterface::dim
-  virtual int dim() const
+  virtual unsigned int dim() const
   {
     return BaseType::size();
   }
@@ -103,7 +103,7 @@ public:
     if (dim() != x->dim())
       DUNE_PYMOR_THROW(Exception::sizes_do_not_match,
                        "size of x (" << x->dim() << ") does not match size of this (" << dim() << ")!");
-    for (int ii = 0; ii < dim(); ++ii)
+    for (size_t ii = 0; ii < dim(); ++ii)
       BaseType::operator[](ii) += alpha * x->operator[](ii);
   } // ... axpy(...)
 
@@ -115,7 +115,7 @@ public:
       DUNE_PYMOR_THROW(Exception::sizes_do_not_match,
                        "size of other (" << other->dim() << ") does not match size of this (" << dim() << ")!");
     double result = 0;
-    for (int ii = 0; ii < dim(); ++ii)
+    for (size_t ii = 0; ii < dim(); ++ii)
       result += BaseType::operator[](ii) * other->operator[](ii);
     return result;
   } // ... dot(...)
@@ -142,7 +142,7 @@ public:
   virtual std::vector< double > components(const std::vector< int >& component_indices) const
     throw (Exception::sizes_do_not_match, Exception::index_out_of_range)
   {
-    if (int(component_indices.size()) > dim())
+    if (component_indices.size() > dim())
       DUNE_PYMOR_THROW(Exception::sizes_do_not_match,
                        "size of component_indices (" << component_indices.size() << ") is larger than this (" << dim() << ")!");
     std::vector< double > values(component_indices.size(), 0);
@@ -151,7 +151,7 @@ public:
       if (component < 0)
         DUNE_PYMOR_THROW(Exception::index_out_of_range,
                          "component_indices[" << ii << "] is negative (" << component << ")!");
-      if (component >= dim())
+      if (component >= int(dim()))
         DUNE_PYMOR_THROW(Exception::index_out_of_range,
                          "component_indices[" << ii << "] is too large for this (" << dim() << ")!");
       values[ii] = BaseType::operator[](component);
@@ -163,7 +163,7 @@ public:
   virtual std::vector< double > amax() const
   {
     std::vector< double > result(2, 0.0);
-    for (int ii = 0; ii < dim(); ++ii) {
+    for (size_t ii = 0; ii < dim(); ++ii) {
       const double value = std::abs(BaseType::operator[](ii));
       if (value > result[1]) {
         result[0] = ii;
@@ -181,7 +181,7 @@ public:
       DUNE_PYMOR_THROW(Exception::sizes_do_not_match,
                        "size of other (" << other->dim() << ") does not match size of this (" << dim() << ")!");
     ThisType* result = create(other->dim());
-    for (int ii = 0; ii < dim(); ++ii)
+    for (size_t ii = 0; ii < dim(); ++ii)
       result->operator[](ii) = BaseType::operator[](ii) + other->operator[](ii);
     return result;
   } // ... add(...)
@@ -193,7 +193,7 @@ public:
     if (dim() != other->dim())
       DUNE_PYMOR_THROW(Exception::sizes_do_not_match,
                        "size of other (" << other->dim() << ") does not match size of this (" << dim() << ")!");
-    for (int ii = 0; ii < dim(); ++ii)
+    for (size_t ii = 0; ii < dim(); ++ii)
       BaseType::operator[](ii) += other->operator[](ii);
   } // ... iadd(...)
 
@@ -205,7 +205,7 @@ public:
       DUNE_PYMOR_THROW(Exception::sizes_do_not_match,
                        "size of other (" << other->dim() << ") does not match size of this (" << dim() << ")!");
     ThisType* result = create(other->dim());
-    for (int ii = 0; ii < dim(); ++ii)
+    for (size_t ii = 0; ii < dim(); ++ii)
       result->operator[](ii) = BaseType::operator[](ii) - other->operator[](ii);
     return result;
   } // ... sub(...)
@@ -217,7 +217,7 @@ public:
     if (dim() != other->dim())
       DUNE_PYMOR_THROW(Exception::sizes_do_not_match,
                        "size of other (" << other->dim() << ") does not match size of this (" << dim() << ")!");
-    for (int ii = 0; ii < dim(); ++ii)
+    for (size_t ii = 0; ii < dim(); ++ii)
       BaseType::operator[](ii) -= other->operator[](ii);
   } // ... isub(...)
 }; // class DuneDynamicVector
